+++ CS.Changelog.Console/Options.cs
@@ -38,10 +38,10 @@ namespace CS.Changelog.Console
     ///                            In order to make an incremental change log since a specified tag. If no tag is specified auto-detects the last release tag.
     ///                            If a tag is specified, overrides option --full
     ///                               
-    ///  --issueformat             (Default: <see cref="Exporters.ExportOptions.IssueNumberRegexDefault"/>)
+    ///  --issueformat             (Default: <see cref="Exporters.BaseOptions.IssueNumberRegexDefault"/>)
     ///                            Expression for recognizing issue numbers
     ///                               
-    ///  --issuetrackerurl         (Default: https://project.cs.nl/issue/{0})
+    ///  --issuetrackerurl         (Default: https://projects.cohelion.com/issue/{0})
     ///                            Url for recognizing issue numbers. '{0}' will be substituted with issue number
     ///                               
     ///  --repositoryurl           Url for showing commit details
@@ -212,7 +212,7 @@ namespace CS.Changelog.Console
         /// <seealso cref="IssueNumberRegex"/> 
         [Option(
              "issuetrackerurl",
-            Default = @"https://project.cs.nl/issue/{0}",
+            Default = @"https://projects.cohelion.com/issue/{0}",
             HelpText = "Url for recognizing issue numbers. '{0}' will be substituted with issue number")]
         [SuppressMessage("Design", "CA1056:Uri properties should not be strings", Justification = "DTO for commandline input cannot accepts urls")]
         public string IssueTrackerUrl { get; set; }