+++ CS.Changelog/ChangeLog.cs
@@ -1,5 +1,6 @@
 ﻿using CS.Changelog.Exporters;
 using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 
@@ -10,20 +11,18 @@ namespace CS.Changelog
 	/// </summary>
 	/// <seealso cref="List{T}" />
 	[JsonObject(MemberSerialization.OptIn)]
-    [SuppressMessage("Naming", "CA1710:Identifiers should have correct suffix", Justification = "Name implies collection")]
-    [SuppressMessage("Naming", "CA1724:Type names should not match namespaces", Justification = "Pending better naming")]
-    public class ChangeLog : List<ChangeSet>
+	[SuppressMessage("Naming", "CA1724:Type names should not match namespaces", Justification = "Pending better naming")]
+	public class ChangeLog : List<ChangeSet>
 	{
 		/// <summary>Gets the repository URL. Allows serialized, unformatted exports to create links to commit details.</summary>
 		/// <value>The repository URL.</value>
 		/// <seealso cref="ExportOptions.RepositoryUrl"/>
 		[JsonProperty(Order = 1)]
-        [SuppressMessage("Design", "CA1056:Uri properties should not be strings", Justification = "<Pending>")]
-        public string RepositoryUrl { get; set; }
+		public Uri RepositoryUrl { get; set; }
 
 		/// <summary>Gets the issue number regex. Allows serialized, unformatted exports to recognize references to issues.</summary>
 		/// <value>The issue number regex.</value>
-		/// <seealso cref="ExportOptions.IssueNumberRegex"/>
+		/// <seealso cref="BaseOptions.IssueNumberRegex"/>
 		[JsonProperty(Order = 3)]
 		public string IssueNumberRegex { get; set; }
 
@@ -31,19 +30,15 @@ namespace CS.Changelog
 		/// <value>The issue tracker URL.</value>
 		/// <seealso cref="ExportOptions.IssueTrackerUrl"/>
 		[JsonProperty(Order = 2)]
-        [SuppressMessage("Design", "CA1056:Uri properties should not be strings", Justification = "<Pending>")]
-        public string IssueTrackerUrl { get; set; }
+		public Uri IssueTrackerUrl { get; set; }
 
 		/// <summary>Gets or sets the change sets. For serialization purposes only</summary>
 		/// <value>The change sets.</value>
 		[JsonProperty(Order = 4)]
-		private IEnumerable<ChangeSet> ChangeSets {
-			get{
-				return ToArray();
-			}
-			set {
-				AddRange(value);
-			}
+		private IEnumerable<ChangeSet> ChangeSets
+		{
+			get => ToArray();
+			set => AddRange(value);
 		}
 	}
 }